make generic Sidenav, add items for publishing&view in LayoutWithRouting
