--- src/components/PublishingSideNav/index.ts
@@ -1,3 +1,0 @@
-import PublishingSideNav from './PublishingSideNav'
-
-export default PublishingSideNav
