+++ src/components/Sidenav/FollowedChannels.style.ts
@@ -2,7 +2,7 @@ import { sizes, colors, typography } from '@/shared/theme'
 import styled from '@emotion/styled'
 import ChannelLink from '../ChannelLink'
 import Text from '../../shared/components/Text'
-import { NAVBAR_LEFT_PADDING, EXPANDED_SIDENAVBAR_WIDTH } from './ViewerSideNav.style'
+import { NAVBAR_LEFT_PADDING, EXPANDED_SIDENAVBAR_WIDTH } from './Sidenav.style'
 
 export const ChannelsTitle = styled(Text)`
   margin-top: ${sizes(6)};
