+++ src/components/Sidenav/index.ts
@@ -1,0 +1,5 @@
+import Sidenav, { NavItem, NavItemType } from './Sidenav'
+import { EXPANDED_SIDENAVBAR_WIDTH } from './Sidenav.style'
+
+export { Sidenav as default, EXPANDED_SIDENAVBAR_WIDTH, NavItem }
+export type { NavItemType }
