--- src/components/ViewerSideNav/index.ts
@@ -1,5 +1,0 @@
-import SideNav, { NavItem, NavItemType } from './ViewerSideNav'
-import { EXPANDED_SIDENAVBAR_WIDTH } from './ViewerSideNav.style'
-
-export { SideNav as default, EXPANDED_SIDENAVBAR_WIDTH, NavItem }
-export type { NavItemType }
