+++ src/components/index.ts
@@ -14,8 +14,7 @@ export { default as ErrorFallback } from './ErrorFallback'
 export { default as ChannelLink } from './ChannelLink'
 export { default as BackgroundPattern } from './BackgroundPattern'
 export { InfiniteVideoGrid, InfiniteChannelGrid } from './InfiniteGrids'
-export { default as ViewerSideNav, EXPANDED_SIDENAVBAR_WIDTH, NavItem } from './ViewerSideNav'
-export { default as PublishingSideNav } from './PublishingSideNav'
+export { default as Sidenav, EXPANDED_SIDENAVBAR_WIDTH, NavItem } from './Sidenav'
 export { default as InterruptedVideosGallery } from './InterruptedVideosGallery'
 export { default as ViewWrapper } from './ViewWrapper'
 export { default as Portal } from './Portal'
